improved & bug fixed

+++ chkGPXfiles.php
@@ -5,13 +5,17 @@
 Если он оканчивается на </trkpt> - в файл дописывается необходимое для получения корректного gpx
 Если файл заканчивается на \n - считается, что файл завершён и корректен.
 
-Если нужно завершить и текущий трек (при обработке треков после всего) - нужно вызвать с параметром all
+Текущим треком считается последний (первый) незавершённый трек. Он не завершается, как признак
+того, что нужно запустить запись трека.
+Если нужно завершить и текущий трек (при обработке треков после всего) - нужно вызвать 
+с параметром all. Однако, если будет обнаружен работающий $gpxlogger, то файл, который он пишет,
+завершён не будет.
 
 cd /www-data/www/GaladrielMap/
 /usr/bin/php-cli chkGPXfiles.php all
 
 */
-chdir(__DIR__); // задаем директорию выполнение скрипта
+chdir(__DIR__); // задаем каталог выполнение скрипта
 
 require_once('fcommon.php');
 require('params.php'); 	// пути и параметры
@@ -27,6 +31,10 @@ if($argv) { 	// cli
 else {	// http
 	$allMaps = $_REQUEST['allMaps'];
 }
+// Определяем, работает ли сейчас запись трека
+$recordedTrackName=gpxloggerRun(true);	// поскольку неизвестно, что содержится в строке global $gpxlogger, нужно определить PID именно этого процесса
+$recordedTrackName = substr(end(explode('/',$recordedTrackName)),0,-4);	// может быть, пора уже исправить basename, а не вводить новые идиотские конструкции в язык?
+//echo "\nchkGPXfiles.php start recordedTrackName=$recordedTrackName;\n";
 
 // Получаем список имён треков
 //echo "trackDir=$trackDir; \n";
@@ -44,11 +52,16 @@ if(!$allMaps) {
 	else unset($trackInfo[count( $trackInfo )-1]);
 }
 //echo "trackInfo:<pre>"; print_r($trackInfo); echo "</pre>\n";
+$reportFileName = $trackDir."/chkGPXfiles_report";
+file_put_contents($reportFileName,date('Y-m-d_His')." Now recorded track - $recordedTrackName; \n");
 foreach($trackInfo as $trk){
+	if($trk==$recordedTrackName) continue;	// ничего не делаем с пишущимся сейчас файлом
 	echo "\n<br> Поехал файл $trk.gpx размером " . filesize( "$trackDir/$trk.gpx" ) . "\n";
+	file_put_contents($reportFileName,"Prepared file $trk.gpx \n",FILE_APPEND);
 	if( filesize( "$trackDir/$trk.gpx" ) <= 573 ) {
 		if(unlink( "$trackDir/$trk.gpx" ) !== FALSE)	{	
 			echo "удалён короткий файл $trackDir/$trk.gpx \n"; 	// незавершённый файл с одной точкой
+			file_put_contents($reportFileName,"Deleted short file $trk.gpx \n",FILE_APPEND);
 			continue;
 		}
 	}
@@ -61,10 +74,12 @@ foreach($trackInfo as $trk){
 		if($lastStr == '</trkpt>') {  	echo "трек готов к завершению\n";
 		    	if(file_put_contents( "$trackDir/$trk.gpx", "\n </trkseg>\n </trk>\n</gpx>", FILE_APPEND ) !== FALSE) {
 		    		echo "завершён файл $trackDir/$trk.gpx \n";
+					file_put_contents($reportFileName,"\tFinished file $trk.gpx \n",FILE_APPEND);
 		    		break;
 		    	}
 		}
 		else {  	echo "трек не завершён, обрезаем файл на $lastStrLen байт\n";
+			file_put_contents($reportFileName,"The track is not completed, will trim the file by $lastStrLen bytes \n",FILE_APPEND);
 			$h = fopen("$trackDir/$trk.gpx", 'r+');
 			$trkFileSize -= $lastStrLen; 	// невозможно каждый раз определять размер файла, потому что кэш
 			ftruncate($h, $trkFileSize);
